change type of column in migration insurance immovable, float to integer, before rails db:migrate

+++ db/migrate/20200211163503_create_insurance_immovables.rb
@@ -1,7 +1,7 @@
 class CreateInsuranceImmovables < ActiveRecord::Migration[6.0]
   def change
     create_table :insurance_immovables do |t|
-      t.float :building_height
+      t.integer :building_height
       t.integer :number_of_floors
     end
   end
